Use value objects #321 Refactor import path
