+++ src/routes/translate/+page.svelte
@@ -14,10 +14,10 @@
 	import { DeleteTextApi } from '$lib/text/delete_text_api'
 	import { TextsApi } from '$lib/text/texts_api'
 	import { LocaleSelectElement } from '$lib/view/locale_select_element'
-	import type { PageData } from '.svelte-kit/types/src/routes/$types'
 	import type { Locale, Text } from '@prisma/client'
 	import { onMount } from 'svelte'
 	import { locale, waitLocale, _ } from 'svelte-i18n'
+	import type { PageData } from './$types'
 
 	export let data: PageData
